+++ src/components/features/productItem/ProductItem.tsx
@@ -1,11 +1,11 @@
 import * as S from './ProductItem.styles';
-import Button from '../@common/button/Button';
-import type { ProductItemType } from '../../types/data';
+import Button from '../../@common/button/Button';
+import type { ProductItemType } from '../../../types/data';
 import AddShoppingCartIcon from '/public/icon/add-shopping-cart.svg';
-import CounterControl from '../@common/counterControl/CounterControl';
-import { CLIENT_ERROR_MESSAGE } from '../../constants/errorMessages';
-import useCartCount from '../../hooks/useCartCount';
-import { handleImageError } from '../../util/handleImageError';
+import CounterControl from '../../@common/counterControl/CounterControl';
+import { CLIENT_ERROR_MESSAGE } from '../../../constants/errorMessages';
+import useCartCount from '../../../hooks/features/useCartCount';
+import { handleImageError } from '../../../util/handleImageError';
 interface ProductItemProps {
   cartInCount: number;
   product: ProductItemType;
